Move remaining fs watching tests to AndroidIncrementalExecutionPerformanceTest
